Split off Pool module
